Correct response code for mock batch picture fetch
Getting a picture without the `redirect=false` parameter returns a 302
redirect. In the batch API, it renders the response inline, telling us
the http status code that the request would have returned. This was
mistakenly listed in the mock as a '200' instead of a '302'.

You can verify this response from the command line (this one fetches
Zuck's profile photo):

  curl \
    -F 'access_token=<REDACTED>' \
    -F 'batch=[{"method":"GET", "relative_url":"4/picture"}]' \
    https://graph.facebook.com